requirements.txt
gensim==3.8.1
flask==1.1.1
pillow==6.2.1
keras-applications==1.0.8
keras-preprocessing==1.1.0
numpy==1.17.4
pandas==0.25.3
nltk==3.4.5
scikit-learn==0.22
tensorflow==1.14.0
scipy==1.3.3
requests==2.22.0
urllib3==1.25.7
gunicorn==20.0.4